Move QuestionMenu out Question because it is a visual component and not part of Question model
